requirements.txt
annotated-types==0.6.0
anyio==4.4.0
argcomplete==3.3.0
asgiref==3.8.1
astroid==3.2.4
attrs==23.2.0
autopep8==2.1.0
backoff==2.2.1
boto3==1.34.105
botocore==1.34.105
build==1.2.1
CacheControl==0.14.0
cairocffi==1.7.0
CairoSVG==2.7.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crashtest==0.4.1
cryptography==42.0.7
cssselect2==0.7.0
DateTime==5.5
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
Faker==25.2.0
fastapi==0.111.0
fastapi-cli==0.0.3
fastjsonschema==2.19.1
filelock==3.14.0
finite-state-machine==0.5.0
ghostscript==0.7
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
injector==0.21.0
installer==0.7.0
isort==5.13.2
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
JPype1==1.5.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
linecache2==1.0.0
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
numpy==1.26.4
openpyxl==3.1.2
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pillow==10.3.0
pip==24.0
pipenv==2023.12.1
pipx==1.5.0
pkginfo==1.10.0
platformdirs==4.2.2
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
protobuf==4.25.3
ptyprocess==0.7.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.3
PyMySQL==1.1.1
pypdf==4.2.0
pypdfium2==4.30.0
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyrsistent==0.20.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
rapidfuzz==3.9.0
referencing==0.35.1
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.1
schedule==1.2.2
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
svgelements==1.9.6
tabula-py==2.9.1
tabulate==0.9.0
thrift==0.20.0
tinycss2==1.3.0
tomli==2.0.1
tomlkit==0.12.5
traceback2==1.4.0
trove-classifiers==2024.4.10
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
userpath==1.9.2
uvicorn==0.29.0
virtualenv==20.26.6
virtualenv-clone==0.5.7
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
zipp==3.19.1
zope.interface==6.3
fuzzywuzzy==0.18.0

